fix(overview): Fixes issue with overview showing wrong "to place"
Fixes #178

There was an issue with async handling requests and race conditioning
where if you searched for something before previous request was
finished and the next request returned no result, it would show
previous trip, but new to place. This is now fixed.

This also technically improves performance somewhat, as we cancel
previous requests as soon as we can, preventing further work and
processing. These improvements were also added to search modal