feat: Add component for highlighting news (#635) (#741)
Add an additional parameter 'highlightComponent' to the
DisapperingHeader, which may be used for showing highlighted text and/or
images.

This new component is used on the search assistant screen where a news
banner is shown. The text and link for this news banner is fetched from
the remote config in Firebase.