feat: Fetch reference data from RemoteConfig
Instead of retrieving reference data by fetching them with http requests
from the referene data services, the reference data is now fetched from
the RemoteConfig. In addition default reference data is bundled with the
app, as backup for situations where the app is unable to retrieve it
from RemoteConfig. Using RemoteConfig also gives out-of-the-box caching
for the user.