feat: Show ref data names in correct langauge
The reference data (UserProfiles, TariffZones, PreassignedFareProducts)
format support multiple languages by providing names and other texts in
a type which has a field 'lang' specifying language and a field 'value'
containing the text in that language. In our code this is typed as a
'LanguageAndText'.

This commit add utility functions for selecting the correct text based
on the current language.