refactor: Standardize header button types (#908)
The screen header mostly had the same functionality for its buttons. The
left buttons were either 'Back', 'Close', 'Cancel' or 'Home', and the
right button was always 'Chat'.

These button types are now standardized as a part of the ScreenHeader
props. Meaning the icon/text, onPress and accessibility is defined in
the ScreenHeader, making it not necessary to duplicate this all over the
app.

Also we were not consistent with usage of text or icons for these header
buttons. It will be easier to be consistent with this new
standardization. After this change mainly text will be used (except for
'Home' and 'Chat), since this is more inline with design guidelines, and
it will give a better differentiation between 'Close' and 'Cancel'.

A custom type may be added later for specifying buttons that is not
included in these standardized types, but as of now I found it not
necessary to add.