feat: Add underline for resend code button (#993)
Added the underline by using TouchableOpacity and ThemeText instead of
using button. Modifying button by specifying text type, or adding a mode
"link" would have given more reusability, but this would be adding modes
that is outside the design system.

Also removed the possibility to specify text types for buttons. The text
types should be whatever was intended for that button mode in the design
system.