feat: Show alerts in app (#1006)
Alerts may be configured in FireStore and will immidiately be shown in
the app. The alert are are configured with a context field, which
specifies if the context is travel, ticketing or general. If the context
is travel the alert is shown on the Assistant and Nearby screen, if
ticketing the alert is shown on Ticket screen (including Overview and
Confirmation screen), and if general the alert will be shown on all
mentioned screens.

Also made some changes to the MessageBox representation. Now the message
box background is a solid background color, to make it not be affected
by the underlying background color.