feat: Set background color for React navigation
During transitions the generic white background from React naviation
some times "leaked" through, giving a noticable flash to the user,
especially in dark mode. We got report from a user in Intercom that
this was unpleasant.

The best fix would be to remove these "leaks" of the background color
all together, but I could not find a good solution for this. As such
this was fixed by setting the background color for React navigation to
be 'background_1' from the AtB theme. This makes the background light
gray in light mode and dark gray in dark mode, making the background
leek less unpleasant.