feat: List tokens as source for inspectable token
Mainly use list tokens as the source for which token is inspectable.
This gives less possibility for the firestore-tcard-value, local token
state, and remote tokens, to become out of sync.

Some places the local firestore-tcard-value must still be used to
support users that are not yet enrolled with mobile token.

Some texts are still missing, and are marked with TODO.