refactor: Fix favorite toggle not working on quay
In the quay view the favorite toggle was not working. This was because
the outdated state-param 'showOnlyFavoriteDepartures' were used instead
of the updated action-param with the same name.

To clean up the logic I removed the field 'showOnlyFavoriteDepartures'
from the state, since it was not actually necessary to have there. The
logic can be just based on the field 'favoriteDepartures', which
already is undefined is 'showOnlyFavoriteDepartures' is false.