fix: Show realtime even though before aimed time
When the exexcted time was before aimed time, the expected time was
shown. I am not sure if this was intentional or not. But when searching
for trips all legs of the trip is based on the realtime data, so we need
to show the realtime value for all legs.