build: Fix grep for latest release tag (#3093)
Negative lookahead is not supported in [POSIX Extented Regular Expressions](https://en.wikibooks.org/wiki/Regular_Expressions/POSIX-Extended_Regular_Expressions),
which is what's used with `grep -E`. A possible solution to this is to
chain two `grep` expressions, where one of them inverts the match. As it
is now it matches the organization ID wherever it is in the tag, but I
don't see that as an issue.

I also went ahead and added `fram` to the expression, since that will
potentialle be needed in the near future.