feat: Make boarding info less intrusive
Instead of using message boxes, just have the boarding info as
secondary text under the quay name. Also skip redundant boarding
info on first and last stop of service journey.

Also fixed som padding on the TripRow for situation message, so it
is a little closer to the quay name above.