refactor: Only proved quay id to PlaceScreen (#3126)
The parameter selectedQuay to the PlaceScreen was on object. This
was not beneficial as it is better that the PlaceScreen itself is
in charge of what data is shown. Now when just passing the id, then
the quay data from the stop-departures query are shown, and not the
quay data available at the screen where it linked to the PlaceScreen.