refactor: Support for AbortController in use-pollable-resource
The reason for this is so the hook doesn't return outdated data. If
something like the currently selected quay is changed on the screen
using use-pollable-resource, then a request in progress for the
previously selected quay should be aborted and not update state.

This functionality may be utilized by passing a data-retrieval
function that takes in an AbortSignal as argument.