refactor: 'compact' as own field on button
Type 'compact' mainly decided the internal padding of the button,
but it also made the button behave as it had type 'inline'. The
compact mode was extracted from the type field into its own field,
so it is possible to specify compact mode both when the button
type is 'inline' or 'block'.