refactor: Add ToggleSectionItem
The old 'switch' mode on ActionSectionItem had some screen reader
challenges. This is fixed in this commit, alongside some
restructuring.

- Created a new ToggleSectionItem instead of using
ActionSectionItem.
- Renamed FixedSwitch to Toggle. This is because toggle is used as
the word for the component in design-sketches/Figma, and it is
slightly more intuitive.
- Parameter names are now 'value' and 'onValueChange' so it is the
same parameters as using the Toggle (previously FixedSwitch)
directly.
- When using screen reader the whole section item is now selected,
and some extra measures has been added to ensure the accessibility
state is read correctly.