feat: Add visual representation of applied filters (#3201)
When a filter is applied to a travel search we show a button (of
type pill) representing the applied filter. The button is clickable,
and if clicked will remove the filter.

In addition there is a small notification icon added to the filter
button icon if any filters are applied. The same is done to the
travel time icon if the time is something else than 'now'.