feat: Add flex discount info
When the field `flexDiscountLadder` is returned in the offer
response, the flex discount info is shown. Although we only have
flex discount for adult now, the app supports discounts on multiple
user profile types.

The expandable with the info is as of now always expanded at start
of the purchase process. Might do something about this later, like
storing the state of the expandable on the device storage.

The description text has a reference to a specific zone and number
of days. Because of this it is made overrideable by specifying a
'discountInfo' under 'appTexts' in Firestore Configuration.

Also a minor change to the ExpandableSectionItem which makes the
full height/width of the expandable heading be focusable for the
screen reader.