fix: Stale animation caused wrong content on purchase tab
There is a bug with the react-navigation material top tab library
on Android, where after navigating to it from other places in the
app (and also through deep linking to it) the animation is not
triggered when changing tab, meaning the selected tab is correct,
but the content shown is actually from the previously selected tab.

I did not find any other easy enough way to fix this except
disabling animation on the purchase tabs.