fix: Duplicate key on recent fare contracts
The react component key for recent fare contracts only used fare
product and travellers, and not zones. Which ment that we got a
duplicate key error if a user had two recent fare contracts
with same product and travellers, but different zones.

When doing this the logic for the key ended up being almost a
duplicate of checking for unique fare contracts. So now we use the
underlying fields for the recent fare contract to create an id
which can be used both as react key and for filtering out
duplicates.