feat: Trip search fields now has parallax scroll (#3509)
The trip search fields will now disappear with a parallax effect
when scrolling. This is done to give more space for the trip search
results.

Also made some changes:
- Fixed a bug in ParallaxScroll (old ContentWithDisappearingHeader)
that made the scroll be buggy on iOS when using RefreshControl.
This is now fixed by using padding for offsetting the children,
instead of using contentInset/contentOffset. This was as it was
done for Android
- Some renaming to hopefully easier to more precise names.