feat: Add onboarding for mobile token without travelcard
FRAM is not going to use or refer to travelcards in the app before we are all over on a new system and not using both an old and a new one. We are still going to utilize mobile token and the advantages that come with that, but we don't want to show information about travelcards. As a result of that I duplicated the existing onboarding for mobile token and called it MobileTokenWithoutTravelcardOnboarding, and removed references to travelcard. The SelectTravelTokenScreenComponent I did not duplicate, but I did a check for `disable_travelcard` (which is a new variable in remote config to control this) to see if the choice for travelcard should be hidden, in addition to using the correct translations.

The illustrations are updated with correct colors in https://github.com/AtB-AS/design-system/pull/131