refactor: Restrict imports from stacks-hierarchy (#3670)
Components and modules outside of stacks-hierarchy shouldn't import
things from stacks-hierarchy. It should be the other way around,
that screens in stacks-hierachy imports components and modules from
the outside.

Also fixed the places where this rule broken.

One found exception to the rule is importing screen params, which
are necessary some places to enforce correct params on navigational
onPress handling. In these few cases the eslint rule may just be
disabled for the line.