fix: Set customerAccountId on reserveOffer
In march, when implementing sub accounts, it was added a field
´customer_account_id´ in the reserve offer endpoint, which also
was added to the reservations in Firestore.

It was added code in the app to check that a reservation has the
correct customer account id, however it was not added code to
set the customer account id when reserving a offer. As such no
reservations created from the app showed up in the app.