refactor: Fetch harbors with react-query
Because we are also going to show harbor names when showing boat
fare contracts, we needed to load the harbor names in a way that
is accessible throughout the app (shared state).

This could be done by doing it in for example the TicketingContext,
or maybe adding a new HarborContext. Also we need some retry logic
in the context we put it in, so the harbors are actually loaded
when we want to show them. Also we might wan't to be update the
data at some point, so we don't have several days old stale data in
the app.

At this point I felt it was about time we introduced a query
library instead of adding another pretty complex custom data fetch
variation in the app. We already have variations using
'useInterval' and 'usePollableResource' around the app, but none
of them could be used out of the box for our use case.

I talked with Hanne which had previously used react-query from
TanStack, and she had good experience with it on the projects she
used it on. With react-query we get pretty much out-of-the-box:

- Support for shared queries
- Retry logic
- Stale data logic
- And more...

In general it seems like a really good addition to the app, and will
probably also solve some of our existing use cases quite nicely,
with less code.