feat: Remove Remote Config force refresh
The force refresh of RemoteConfig should no longer be necessary, as
it was mainly for development before we got the button for
refreshing it in the debug menu.

In addition the remote config force refresh created a full remount
on every app startut, creating significant wait time with just
a blank screen.