refactor: Auth flow
Not really changed much in the auth flow, just refactored so the
details are more optional to read into. Previously looking into
the AuthContext could feel a little overwhelming.

The main change of this refactor is that the AuthContext has an
AuthStatus which may be 'loading', 'creating-account, 'finished',
or 'error'.