fix: Remove spinner stuck fix
Adding/removing RefreshControl based on focus gives odd behaviour
on Android where the scroll offset is lost after navigating.

The adding/removing of RefreshControl was added to fix a stuck
spinner on iOS issue, but this doesn't seem to still happen after
updating react-native versions. So the fix for now is to remove
the stuck spinner fix.