refactor: Make translate func a stable reference (#4025)
The translate function (t) from lobot was not a stable reference,
and recreated on every render. This ment it could not be added to
dependency lists without causing the hooks to be run too often.

In this PR we wrap the translate function with 'useCallback', and
add the 't' function to the dependency lists.