fix: Render crash on TravelTokenBox
React components should be created like normal javascript function,
as that bypasses the rules and optimizations of React. The
ErrorMessages component here did not work as intended, as it made
the app crash with 'Rendered fewer hooks than expected' since the
hooks inside the component was rendered conditionally.

The solution now was inlining the error  message boxes into the
TravelTokenBox component.