fix: Not move to antarctica when opening filters sheet
The reason for the bug was that the Map tried to show a walking
path, but since the filters sheet uses so much of the screen it
wasn't enough space to show to walk path, even if it zoomed all
the way out (showing Antarctica).

The fix is that when the filter sheet is opened, we needed to
reset the map state that said that the departures sheet and the
walk path should be visible.

One drawback of this solution is that there is no animation when
switching from the departures sheet to the filters sheet, but I
couldn't get that to work without causing havoc.