feat: Use native user location pin
The non-native user location pin has an onPress handler which is
hard to circumvent, and the hitbox is quite wide. This means when
the user location pin was close to a entity that should be
clickable (like bus stop, bike station, etc.) the click was not
registered by the map.

To fix this we change to the native location pin. The native
location pin is not interactable, which is kinda exactly what we
want here, as we don't want to register clicks on the location pin.

Another minor drawback is that it is not as pretty as the
non-native pin, but this was ok by designers.