fix: Bottom sheet title focus race condition
Bottom sheet title were always focused, so when another ref was
added to the sheet content it was a race condition which ref was
focused by the screen reader.

Added a parameter to the BottomSheetContainer to specify whether
the title should be focused or not, and fixed up some of the bottom
sheets to use this parameter correctly.