feat: Live position in map not reliant on vehicle position (#4203)
We checked on whether the estimated call had realtime data to
determine whether to show live bus position or not, but it turns
out a vehicle may have live position even if it doesn't have
realtime data.

Changed the logic to be based on departure time instead. If there
is less than 10 minutes until the service journey starts, or it has
started, then we will start to fetch vehicle live position.