feat: Add BorderedInfoBox component
Adds a new component to the src and storybook. This is the bordered
info box introduced in the on behalf of product initiative. Pretty
basic component, where the main cavaet is that the border is based
on the text color with 0.1 opacity. This was since we didn't have
an appropriate color to use in the design-system, and we didn't
want to introduced a new one just for this use case.