feat: Not navigate after login (#4257)
Two main changes in this PR:
- We complete onboarding before the last step of the authentication
process, like sending in the confirmation code. This is so we don't
end up in a race condition between Firebase Auth switching user and
setting the onboarded flag to true.
- We do not navigate anywhere after login, as changing user is
basically a reload of the app, and we end up back at the starting
screen no matter what.