refactor: Booking components not rely on Leg (#4319)
To also show booking info on departures screen the component and
utils can't take in leg, as for the departures screen we don't have
legs, but estimated calls.

This is purely a refactor to be able to use booking components on
departures, and should be no functional changes for the user. The
refactor is done in its own PR to not bloat the PR where the
booking info is actually added to the departures screens.