refactor: Inverse status comparator logic
Earlier a list sorted with the status comparator ended up ordered
from error to undefined. While it inutively makes sense, it is not
correct as the default sort order is from smaller to higher values,
and since error is the highest criticality it should end up in the
end of the list.

Also the sorting in Javascript doesn't behave well with undefined
values, so the starts comparator no longer supports those. The
values have to be defined.