feat: Add feature flag for server time (#4327)
- Add a feature flag to be able to turn of server time if
necessary. The feature flag has the default value true, and can be
set to false in RemoteConfig.
- Also fixed a bug where if the server clock didn't start as it
should, then the interval wasn't fired. This ment that if
something went wrong with the time-lib, we would never get an
updated time. Now the interval always runs every 2500 ms, but
fallbacks to Date.now() if server time not enabled or the time-lib
isn't started.