refactor: Use react-query for listing tokens (#4386)
This is to enable more fine-grained error handling in mobile
token flow when updating the sdk. When we have decoupled the
different requests they can have their own error handling.

Without this change we would need a stored state specifying where
in the flow the error happened, so the correct step was retried.