refactor: Upgrade mobile token 2.0.1
- Add handling of error resolutions in errors from sdk. As of now
only handling the reset resolution which signals that the sdk is
stuck. The rest of the errors doesn't have any specific handling
for now and are mainly treated as before.
- The new sdk version has the possibility for both local logger
and remote logger. For now the previously used logger is passed in
as local logger, and for remote logger we just send the errors
as notify events to Bugsnag. We will see how this turns out in
Bugsnag and if it needs to be revised.
- Added som more libs in the Posfile, which is required
dependencies for the sdk, even if we don't use them yet.