fix: Not create new mobile token on logout (#4452)
When logging out we delete the mobile token for the user, both
locally and remotely. However there was a race condition where if
the logout took some time, the app managed to create a new token
for the user after deleting the old one.

This is fixed by setting the MobileTokenContext in a 'logging-out'
state, where it will treat mobile tokens as disabled, and thus not
create new ones.