feat: Token timeout even if not enabled
Even if the mobile token handling is never started because of the
user not getting auth status 'authenticated', the mobile token
timeout handler should still run. This means that the user will get
fallback after the timeout, even if there is something wrong with
the id token.

Also reset isTimeout value to false when mobile token no longer in
loading state.