fix: Not resubscribe auth on enable_intercom change (#4584)
Since we now check for enable_intercom before updating metadata,
the enable_intercom value was in the dependency list for the
useEffect setting up the auth description. On app startup the
enable_intercom value is initially false, before chaning to true
after RemoteConfig loads, which ends up triggering an unneccesary
auth resubscribe at startup.

Moved the updating of Intercom metadata outside of the auth
subscription to mitigate this.