feat: Select recipient from a list (#4652)
Added integration with the on-behalf-of-accounts endpoint to
retrieve a recipient list. Since we now have three different
recipient selection mode in the app (list, only phone, phone and
name) the logic in the select recipient screen ended up more
complex and needed a rewrite.

Also removed the feature-toggle now, as the select recipient
feature is now completed enough that it can be released.