feat: Only stop places checkbox in location search
When looking at geocoder request logs, we see that only ~8% of
location searches end up in the user selecting an address. They
instead use my position or a stop place.

To make searching for stop places easier we add a checkbox the user
can check so only stop places are returned when searching for a
location. We will enable this to a subset of the users, and see if
it is a feature which will be used.