fix: Fix double offer search
The ProductSelectionType has one product, but the offer search may
return another product with the same product alias id. This is what
happens for carnet Trondheim-Vanvikan.

The double offer search was because of the ProductSelection was
updated after the offer search results came in. This shouldn't be
done, the selection should still be the same even if the product in
the offer response changes.

Also made som changes to ensure that texts shown in the app is the
texts for offer search result product.