refactor: ThemeIcon undefined svg safeguard
We got some reports in Bugsnag where the app has crashed because
of the svg prop sent to ThemeIcon was undefined. This should really
not happen, but it is a little extreme that the app crashes if it
does. This safeguard is added to prevent app crash, rather there
will be no rendered ThemeIcon.