feat: Add purchase selection module
Add a new module for creation, modification and testing of the
purchase selection domain model. It uses a builder pattern for
creating and modifying a purchase selection, which ensures that
the built purchase selection is valid and referentially stable
when not edited. The builder it self is not referentially stable,
and should not be passed around or used in hooks dependency lists.

The purchase selection module is thoroughly tested, and it is
configured a required coverage of 100% for the purchase selection
module. This is also checked during the app build.