chore: Add eslint rules for not leaking values to jsx (#4872)
In react-native the app crashes if trying to render text outside of
Text component. This has happened in our production app more than
once because of checks on length, and to be fair it will probably
happen more if we don't do anything about it.

This change adds an eslint rule that gives error when using in
combination with && inside jsx.

The eslint rule used is not that popular, but seems well created
and safe to use when reading through its code.