feat: Store only stop places checkbox state on device
This is decided to be in place for next round of testing of the
only stop places checkbox. In last round of testing the usage of
the checkbox was quite high at the start, but then dropped of. The
theory is that users don't use the extra click to check off the
checkbox for every search.