refactor: Add more clear and tested fare contract availability status (#4876)
The logic whether a fare contract is upcoming, valid, expired, etc.,
was a little all over the place and hard to really get a grasp on. This
PR is made as an effort to be more precise on what these terms
actually are, and to make it easier to retrieve fare contracts with
specific availability status.

The new utils and hooks are not taken in use in this PR. I found
that taking them in use gave a lot of changes, but I wanted the PR
to be standalone to make possible for better reviews of the actual
logic, which is important to be correct and understandable. After
this is merged it will be another PR to replace old util functions
with usage of these new ones.