fix: disable fetching of vehicles when out of focus
We're still seeing a lot of requests towards vehicles in 1.60.1, and it
seems it might be caused by requests that are active without focus. E.g.
if the user navigates from trip details to departure details or switches
tab, the trip query is still active and fetching updates every 20s. And
if several screens with vehicles queries are open, several requests are
sent every 20s.