fix: Some date tests failing after new year (#4899)
The tests need to have current year provided programatically to not
need manual fixing every new year.

In addition when adding some more tests it came to light that the
formatting of the day and month differed based on whether it was
the current year or not. The full date formatting was changed to
give the same day/month format no matter if current year or not.