refactor: remove some unused FareContract/TraveRight fields
Some fields that are always present in Firestore, are not returned by
the ticket service. After adding stricter Zod validation of fields, this
caused tickets from the endpoint to fail validation. Since these fields
were not used (in a significant way) in the app, I opted to treat them
as deprecated and remove them from the app entirely.

At the same time, it shows that we should use the ticket service as
source of truth for fare contract and travel right types in the app.